Decompose app entrypoint model (#1371)
Phase 3 file decomposition slice for #1195 / #1145, stacked on #1370.

- Split oversized `app/app.go` (1782 lines) into focused files:
  - `app.go` keeps the small TUI entrypoint.
- `home_model.go` holds state/model construction, layout, focus, and
Init.
  - `home_update.go` holds Bubble Tea Update and key routing.
  - `home_tasks.go` holds task create/trigger handling.
- `home_attach.go` holds attach-overlay detach reset and daemon poll
pause heartbeat.
  - `home_panes.go` holds selection/pane refresh helpers.
  - `messages.go` holds shared tea message structs.
  - `home_view.go` holds errors, confirmations, and View.
- Removed `app/app.go` from `scripts/file-length-allowlist.txt`; the
largest split file is 542 lines.
- Updated `docs/file-length-lint.md` to match the active allowlist.

Pure declaration moves; no behavior changes intended.

- `gofmt`
- `go build ./...`
- `golangci-lint run --fast`
- `scripts/lint-file-length.sh`
- `deadcode -test ./...`
- `make test-container`

Refs #1195 Phase 3.

— Captain Claude
