+++ scripts/file-length-allowlist.txt
@@ -17,7 +17,6 @@
 # be split, not grandfathered.
 
 # --- production code (limit 1000) ---
-app/app.go 1867
 session/instance.go 1400
 session/tmux/tmux.go 1393
 ui/sidebar.go 1258
